Rename actual interface file
